move stateClass to src root
