+++ packages/components/src/TextField/common.ts
@@ -1,20 +1,3 @@
-import { Namespace } from "../constants"
+import { createStateClass } from "../stateClass"
 
-export const COMPONENT_CLASS = `${Namespace}-textfield`
-
-type stateMap = {
-  disabled?: boolean
-  readOnly?: boolean
-  error?: boolean
-  variantLine?: boolean
-}
-// stateClass({ disabled: true })
-// stateClass({ disabled: false, readOnly: true })
-export const stateClass = (stateMap: stateMap) => {
-  let ret = []
-  for (const stateName in stateMap) {
-    if (stateMap[stateName as keyof stateMap])
-      ret.push(`${COMPONENT_CLASS}--${stateName}`)
-  }
-  return ret.join(" ")
-}
+export const [stateClass] = createStateClass("textfield")
